Update KordEx to 1.5.2-RC1, remove unused code, fix thread commands working in normal channels, move ThreadModInviter.kt to moderation
