+++ libs.versions.toml
@@ -2,7 +2,7 @@
 kotlin = "1.6.10"  # Note: Plugin versions must be updated in the settings.gradle.kts too
 
 groovy = "3.0.9"
-kord-extensions = "1.5.2-SNAPSHOT"
+kord-extensions = "1.5.2-RC1"
 logging = "2.1.21"
 logback = "1.2.8"
