+++ src/main/kotlin/net/irisshaders/lilybot/utils/ResponseHelper.kt
@@ -1,7 +1,6 @@
 package net.irisshaders.lilybot.utils
 
 import com.kotlindiscord.kord.extensions.DISCORD_BLACK
-import com.kotlindiscord.kord.extensions.DISCORD_RED
 import com.kotlindiscord.kord.extensions.utils.dm
 import dev.kord.common.Color
 import dev.kord.core.behavior.channel.MessageChannelBehavior
@@ -12,23 +11,6 @@ import dev.kord.rest.builder.message.create.embed
 import kotlinx.datetime.Clock
 
 object ResponseHelper {
-	/**
-	 * This function generates a Failure [embed] for ease of use in other places
-	 * @param channel The channel that the embed will be created in
-	 * @param embedTitle The title of the embed, or null if no title
-	 * @param embedDescription The description of the embed, or null if no description needed
-	 * @return the built-in embed
-	 * @author NoComment1105
-	 */
-	suspend fun failureEmbed(channel: MessageChannelBehavior, embedTitle: String?, embedDescription: String?): Message {
-		return channel.createEmbed {
-			if (embedTitle != null) title = embedTitle
-			if (embedDescription != null) description = embedDescription
-			color = DISCORD_RED
-			timestamp = Clock.System.now()
-		}
-	}
-
     /**
      * Using the provided [channel] an [embed] will be returned, populated by the fields specified by the user
      *
